requirements.txt
appnope==0.1.0
atomicwrites==1.2.1
attrs==18.1.0
boto3==1.9.42
botocore==1.12.42
certifi==2018.8.24
chardet==3.0.4
-e git+https://github.com/uktrade/django-staff-sso-usermodel.git@f3ab9261de3b7a69b733a446a148419e12055216#egg=custom_usermodel
decorator==4.2.1
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1.5
django-admin-list-filter-dropdown==1.0.1
django-autocomplete-light==3.3.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-debug-toolbar==1.10.1
django-environ==0.4.5
django-extensions==2.1.2
django-filter==2.0.0
django-import-export==1.1.0
-e git+https://github.com/uktrade/django-staff-sso-client.git@3927b3e07a05ba77c0d14ea46b8609b136131316#egg=django_staff_sso_client
django-storages==1.7.1
django-tables2==1.21.2
docutils==0.14
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==0.9.0
flake8==3.5.0
flake8-import-order==0.18
get==1.0.3
gunicorn==19.9.0
idna==2.7
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.1
jmespath==0.9.3
mccabe==0.6.1
more-itertools==4.3.0
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.4.9
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.7.1
post==1.0.2
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
public==1.0.3
py==1.6.0
pycodestyle==2.3.1
pydot==1.2.4
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.7.4
pytest-django==3.4.2
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.3
PyYAML==3.13
query-string==1.0.2
raven==6.9.0
request==1.0.2
requests==2.20.0
requests-oauthlib==1.0.0
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
text-unidecode==1.2
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.23
wcwidth==0.1.7
whitenoise==4.1
xlrd==1.1.0
xlwt==1.3.0
